+++ Assignment_3/main.c
@@ -4,15 +4,15 @@
 
 int main() {
     char operator;
-    double firstNumber,secondNumber, sum;
+    float firstNumber,secondNumber, sum;
 
     printf("Enter an operator (+, -, *, /): ");
     scanf("%c", &operator);
 
     printf("Enter two operands: ");
-    scanf("%lf %lf",&firstNumber, &secondNumber);
+    scanf("%f %f",&firstNumber, &secondNumber);
 
     sum=calculate(operator,firstNumber,secondNumber);
-    printf("\n%.1lf %c %.1lf = %.1lf",firstNumber, operator, secondNumber, sum);
+    printf("\n%.1f %c %.1f = %.1f",firstNumber, operator, secondNumber, sum);
     return 0;
 }
